Added ChatMessageProvider
Save every message into a database. This will make it possible to search
for old notifications, and to the history of the last session. History
is no longer lost if the users disconnects.

Currently, the default message buffer size is 100. It may be a good
idea, to increase that (use a new settings key), or to limit by time. It
might also be interesting to limit the chat view to only the most recent
X messages, instead of everything.